Optimizing native integer usage

+++ src/ImageSharp/Common/Helpers/Shuffle/IComponentShuffle.cs
@@ -99,15 +99,15 @@ namespace SixLabors.ImageSharp
         {
             ref uint sBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(source));
             ref uint dBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(dest));
-            nuint n = (nuint)(uint)source.Length / 4;
+            nint n = (nint)source.Length / 4;
 
-            for (nuint i = 0; i < n; i++)
+            for (nint i = 0; i < n; i++)
             {
-                uint packed = Unsafe.Add(ref sBase, (int)i);
+                uint packed = Unsafe.Add(ref sBase, i);
 
                 // packed          = [W Z Y X]
                 // ROTL(8, packed) = [Z Y X W]
-                Unsafe.Add(ref dBase, (int)i) = (packed << 8) | (packed >> 24);
+                Unsafe.Add(ref dBase, i) = (packed << 8) | (packed >> 24);
             }
         }
     }
@@ -125,15 +125,15 @@ namespace SixLabors.ImageSharp
         {
             ref uint sBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(source));
             ref uint dBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(dest));
-            nuint n = (nuint)(uint)source.Length / 4;
+            nint n = (nint)source.Length / 4;
 
-            for (nuint i = 0; i < n; i++)
+            for (nint i = 0; i < n; i++)
             {
-                uint packed = Unsafe.Add(ref sBase, (int)i);
+                uint packed = Unsafe.Add(ref sBase, i);
 
                 // packed              = [W Z Y X]
                 // REVERSE(packedArgb) = [X Y Z W]
-                Unsafe.Add(ref dBase, (int)i) = BinaryPrimitives.ReverseEndianness(packed);
+                Unsafe.Add(ref dBase, i) = BinaryPrimitives.ReverseEndianness(packed);
             }
         }
     }
@@ -151,15 +151,15 @@ namespace SixLabors.ImageSharp
         {
             ref uint sBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(source));
             ref uint dBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(dest));
-            nuint n = (nuint)(uint)source.Length / 4;
+            nint n = (nint)source.Length / 4;
 
-            for (nuint i = 0; i < n; i++)
+            for (nint i = 0; i < n; i++)
             {
-                uint packed = Unsafe.Add(ref sBase, (int)i);
+                uint packed = Unsafe.Add(ref sBase, i);
 
                 // packed              = [W Z Y X]
                 // ROTR(8, packedArgb) = [Y Z W X]
-                Unsafe.Add(ref dBase, (int)i) = (packed >> 8) | (packed << 24);
+                Unsafe.Add(ref dBase, i) = (packed >> 8) | (packed << 24);
             }
         }
     }
@@ -177,11 +177,11 @@ namespace SixLabors.ImageSharp
         {
             ref uint sBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(source));
             ref uint dBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(dest));
-            nuint n = (nuint)(uint)source.Length / 4;
+            nint n = (nint)source.Length / 4;
 
-            for (nuint i = 0; i < n; i++)
+            for (nint i = 0; i < n; i++)
             {
-                uint packed = Unsafe.Add(ref sBase, (int)i);
+                uint packed = Unsafe.Add(ref sBase, i);
 
                 // packed              = [W Z Y X]
                 // tmp1                = [W 0 Y 0]
@@ -192,7 +192,7 @@ namespace SixLabors.ImageSharp
                 uint tmp2 = packed & 0x00FF00FF;
                 uint tmp3 = BitOperations.RotateLeft(tmp2, 16);
 
-                Unsafe.Add(ref dBase, (int)i) = tmp1 + tmp3;
+                Unsafe.Add(ref dBase, i) = tmp1 + tmp3;
             }
         }
     }
@@ -210,11 +210,11 @@ namespace SixLabors.ImageSharp
         {
             ref uint sBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(source));
             ref uint dBase = ref Unsafe.As<byte, uint>(ref MemoryMarshal.GetReference(dest));
-            nuint n = (nuint)(uint)source.Length / 4;
+            nint n = (nint)source.Length / 4;
 
-            for (nuint i = 0; i < n; i++)
+            for (nint i = 0; i < n; i++)
             {
-                uint packed = Unsafe.Add(ref sBase, (int)i);
+                uint packed = Unsafe.Add(ref sBase, i);
 
                 // packed              = [W Z Y X]
                 // tmp1                = [0 Z 0 X]
@@ -225,7 +225,7 @@ namespace SixLabors.ImageSharp
                 uint tmp2 = packed & 0xFF00FF00;
                 uint tmp3 = BitOperations.RotateLeft(tmp2, 16);
 
-                Unsafe.Add(ref dBase, (int)i) = tmp1 + tmp3;
+                Unsafe.Add(ref dBase, i) = tmp1 + tmp3;
             }
         }
     }
